Renaming the Delphi module to DelphiVCL to make room for a DelphiFMX
