+++ Modules/DelphiVCL/uMain.pas
@@ -1,0 +1,52 @@
+unit uMain;
+
+interface
+
+uses PythonEngine;
+
+function PyInit_DelphiVCL: PPyObject; cdecl;
+
+implementation
+
+uses WrapDelphi, WrapDelphiVCL;
+
+var
+  gEngine : TPythonEngine;
+  gModule : TPythonModule;
+  gDelphiWrapper : TPyDelphiWrapper;
+
+function PyInit_DelphiVCL: PPyObject;
+begin
+  try
+    gEngine := TPythonEngine.Create(nil);
+    gEngine.AutoFinalize := False;
+    gEngine.UseLastKnownVersion := False;
+    // Adapt to the desired python version
+    gEngine.RegVersion := '3.9';
+    gEngine.DllName := 'python39.dll';
+
+    gModule := TPythonModule.Create(nil);
+    gModule.Engine := gEngine;
+    gModule.ModuleName := 'DelphiVCL';
+
+    gDelphiWrapper := TPyDelphiWrapper.Create(nil);
+    gDelphiWrapper.Engine := gEngine;
+    gDelphiWrapper.Module := gModule;
+
+    gEngine.LoadDll;
+  except
+  end;
+  Result := gModule.Module;
+end;
+
+initialization
+  gEngine := nil;
+  gModule := nil;
+  gDelphiWrapper := nil;
+finalization
+  gEngine.Free;
+  gModule.Free;
+  gDelphiWrapper.Free;
+end.
+
+
